+++ tests/integration_tests/test_dataset.py
@@ -66,6 +66,28 @@ def test_change_data_size():
         assert next(dataset).size() == result_data_size
     
     dataset.close() 
+
+def test_text_dataset():
+    batch_size = 20
+    block_size = 128
+    num_batches = 10
+    epoch_length = 10
+
+    dataset = bittensor.dataset (
+        _mock = True,
+        batch_size = batch_size, 
+        block_size = block_size,
+        num_batches = num_batches
+    )
     
+    dataloader = dataset.dataloader(epoch_length)
+
+    assert len(dataloader) == epoch_length
+    assert len(dataloader) != len(dataset)
+    assert len(dataset[0]) == block_size
+    assert len(dataloader.dataset) == batch_size * epoch_length
+    
+    dataset.close()
+
 if __name__ == "__main__":
     test_change_data_size()
