+++ tests/integration_tests/test_dendrite.py
@@ -20,6 +20,7 @@ import torch
 import pytest
 import bittensor
 from bittensor._proto.bittensor_pb2 import UnknownException
+from bittensor.utils.test_utils import get_random_unused_port
 from . import constant
 
 wallet = bittensor.wallet.mock()
@@ -275,8 +276,9 @@ def test_successful_synapse():
     def forward_casual_lm_next(inputs_x, synapse, model_output=None):
         return None, None, synapse.nill_forward_response_tensor(inputs_x)
 
+    port = get_random_unused_port()
     axon = bittensor.axon (
-        port = 8096,
+        port = port,
         ip = '0.0.0.0',
         wallet = wallet,
     )
@@ -293,7 +295,7 @@ def test_successful_synapse():
         hotkey = wallet.hotkey.ss58_address,
         ip = '0.0.0.0', 
         ip_type = 4, 
-        port = 8096, 
+        port = port, 
         modality = 0, 
         coldkey = wallet.coldkeypub.ss58_address
     )
@@ -323,8 +325,9 @@ def test_failing_synapse():
     def forward_casual_lm_next(inputs_x, synapse, model_output=None):
         return None, None, synapse.nill_forward_response_tensor(inputs_x)
 
+    port = get_random_unused_port()
     axon = bittensor.axon (
-        port = 8097,
+        port = port,
         ip = '0.0.0.0',
         wallet = wallet,
     )
@@ -341,7 +344,7 @@ def test_failing_synapse():
         hotkey = wallet.hotkey.ss58_address,
         ip = '0.0.0.0', 
         ip_type = 4, 
-        port = 8097, 
+        port = port, 
         modality = 0, 
         coldkey = wallet.coldkeypub.ss58_address
     )
@@ -382,8 +385,9 @@ def test_missing_synapse():
     def forward_casual_lm_next(inputs_x, synapse, model_output=None):
         return None, None, synapse.nill_forward_response_tensor(inputs_x)
 
+    port = get_random_unused_port()
     axon = bittensor.axon (
-        port = 8098,
+        port = port,
         ip = '0.0.0.0',
         wallet = wallet,
     )
@@ -396,7 +400,7 @@ def test_missing_synapse():
         hotkey = wallet.hotkey.ss58_address,
         ip = '0.0.0.0', 
         ip_type = 4, 
-        port = 8098, 
+        port = port, 
         modality = 0, 
         coldkey = wallet.coldkeypub.ss58_address
     )
@@ -440,8 +444,9 @@ def test_dendrite_timeout():
         time.sleep(3)
         return None, None, synapse.nill_forward_response_tensor(inputs_x)
 
+    port = get_random_unused_port()
     axon = bittensor.axon (
-        port = 8098,
+        port = port,
         ip = '0.0.0.0',
         wallet = wallet,
     )
@@ -454,7 +459,7 @@ def test_dendrite_timeout():
         hotkey = wallet.hotkey.ss58_address,
         ip = '0.0.0.0', 
         ip_type = 4, 
-        port = 8098, 
+        port = port, 
         modality = 0, 
         coldkey = wallet.coldkeypub.ss58_address
     )
@@ -481,5 +486,4 @@ def test_clear():
     dataset.close()
     
 if __name__ == "__main__":
-    bittensor.logging(debug = True)
     test_dendrite_timeout()
