+++ tests/unit_tests/bittensor_tests/test_config.py
@@ -21,7 +21,6 @@ import bittensor
 import argparse
 import pytest
 
-bittensor.logging(debug = True)
 
 def test_loaded_config():
     with pytest.raises(NotImplementedError):
